refactor(cli): split output responsibilities
